Prevent merging conference calls hosted on peer device.
With IMS, the "multiparty" bit on an ImsCall is set to "true" when a call
is merged into a conference.  This not only occurs on the device hosting
the conference call, but also on the devices of the callers merged into
the conference.

This CL adds a listener to the ImsCallSessionListener which is used to
communicate a change in the multiparty state of an ImsCallSession to
the ImsCall.  This solves a problem where the RIL knows of the change, but
since there is no callback, the change is not noticed by Telephony until
another call state change occurrs (e.g. holding the call).

Bug: 19478784
Change-Id: I4847ab3b63f6a00a91d1324196f181467d1753a4