Don't limit fullscreen stack window size to parent window size
A previous change limited the size of a window to the parent window
size at max. so that child windows don't extend outside their parent
stack when resized in a multi-window environment. This broke the
wallpaper positioning functionality since the wallpaper is no longer
bigger than it's containing stack so it can't be scrolled. Now, we
only limit the window size to the parent window size if the window
stack is not fullscreen.

Bug: 19434096
Bug: 19225079
Change-Id: I1a8788727e6c4a91da45d8a87850093ef5a24edf