Fix crash during Activity Transition with TextureView.
Bug 20675857

GhostView was synchronously forcing the View's parent to get
the display list when it was created. TextureView's cannot
create a display list until a surface is available and will
crash if it is attempted. This change moves the request to
an invalidate to get it to execute at a later time.

Change-Id: I12afebdfd57eb29c1a35b23a63d045e319547faf