Preserve ResolverDrawerLayout fully collapsed state across relayout
Fix a bug where a relayout that grows the size of ResolverDrawerLayout
content would cause it to get stuck in a half-collapsed state. This
could also lead to janky touch interaction behavior in
onStopNestedScroll as it animates to the fully open/closed state after
clicking in a descendant list.

Bug 21202727

Change-Id: I5f0a8b19fc079c57fcf2be7c6fa7627b1cb12a3c