Wait for wakeup_reasons on first init
Previously if we were initializing the wakeup callback for the first time,
we would read the wakeup_reason file and ignore the contents, sending a
wakeup_reason of "unknown" up to BatteryStats.

Now we initialize the callback and wait on it immediately. Wakeup reasons are reset
when we go into sleep, so when we wakeup, we will always have fresh wakeup reasons.

Bug:21665793

Change-Id: I20832d8a143fc2715915fcecf4bb71980f279440