Revert "Load WebView from one out of a list of packages."
This was not a clean revert!

This reverts commit b3050fc43113acc42781b24f425c467a85f38018.
We essentially only revert the functionality for going through a list of
WebView package names and picking the first compatible one.
Except for that functionality we also fetched the name of the shared
library from a flag in WebView and made some minor refactoring in the
initial commit, these changes have been left alone in this revert.

Bug: 21893371
Change-Id: Idb2539dc33cc5f9e2894ecd665c23573c6cba9f3